IDをintからint64に微修正

+++ internal/models/user.go
@@ -5,6 +5,6 @@ import "github.com/uptrace/bun"
 type User struct {
 	bun.BaseModel `bun:"table:users,alias:u"`
 
-	ID   int    `bun:"id,pk,autoincrement"`
+	ID   int64  `bun:"id,pk,autoincrement"`
 	Name string `bun:"name,notnull"`
 }
